Move out the JS runtime into their own files
