+++ src/webcore/runtime_emscripten.js
@@ -1,0 +1,3 @@
+Module.STDWEB.alloc = _malloc;
+Module.STDWEB.dyncall = Runtime.dynCall;
+Module.STDWEB.utf8_len = lengthBytesUTF8;
